requirements.txt
pytrilogy>=0.0.2.15
google-auth
google-cloud-bigquery
sqlalchemy-bigquery